Read parameters from settings.yml file
In order to facilitate the integration in the RHEL agent,
we can read settings from settings yaml file.

LOCAL_STORAGE_FOLDER has been renamed in DCI_LOCAL_REPO to be consistent with the agent.
For a more friendly API topic and topics within the settings file are converted into topics_names.

Change-Id: Icb0a97bd3fd6a886fe4cfd6fde36e0d245082ca6