Update patch set 1
Patch Set 1: Code-Review-1

I agree there are some issues on the import and implicit relative import are deleted in python 3. Nice catch.

But I'm not sure the pep328 relative import is the solution.

Google style guide write: 

> Do not use relative names in imports. Even if the module is in the same package, use the full package name. This helps prevent unintentionally importing a package twice.

Can we use absolute import instead?

If there is an issue in the NamedTemporaryFile, could we add a test for this one? dci-downloader is tested on py3 and py2. If the CI is green, it's because there is no test covering this.

So instead of the comment, a test will be better

Patch-set: 1
Label: Code-Review=-1