fix backward compatibility issue
in the function _keep_backward_compatibility() when
a key has to be backwarded then the following keys
are not taken into account.

For instance if the settings contains 'topic' and
'download_folder', only the 'topic' key will be
returned.

This patch fixed the issue by returning new_settings
variable at the end of the function.

Change-Id: Ic3b6056df775ec8359f62769bc15ef7c36d69350