Update patch set 3
Patch Set 3: Code-Review-1

(4 comments)

I don't understand, is the API described in the card allows to download multiple components? I think we only want to download a specific component by id. That it.

If we a have a --component param in the cli, we want to get the component and download it no?

Patch-set: 3
Label: Code-Review=-1