main.py: fix job variable scope in try/except
the job variable is assigned in the try block but
also accessed in the except block. If the statement
before the job assignment fail then the except block
will access the job variable which would be not declared.
Causing the error "UnboundLocalError: local variable
'job' referenced before assignment"

This commit fix this issue by assigning the job variable
before the try/except block.

Change-Id: Ie118265a5e0b2a392073baa2fb5db40c0bb51a5d