Update patch set 5
Patch Set 5: Code-Review+2

(1 comment)

Calling the dci-downloader CLI multiple times has the advantage that most of the time it's used from Ansible, and if one component download fail, it would be straightforward to diagnostic the issue. The drawback would be that we will ends with noise "download" jobs. So i'm also fine to download multiple components in one CLI call.

Also, i'm  wondering the following, for example:

If a job needs 3 components for a given topic (3 components type) and the user is specifying one then the job will only be attached to this alone component ? 

Or do we want to add/override it with the latest of each component type needed by the topic?

Patch-set: 5
Reviewer: Gerrit User 29 <29@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=+2