Add file lock to avoid to download the same topic in two processes simultaneously
When we try to download the same topic in two differents processes,
we can encounter issues and race conditions. Now we use one file lock per topic.
If file lock is present we wait 30s and check again.
We do this during 10 hours before we stop waiting.

This one allow complete parallel downloads.

Change-Id: I5c8d450698064a26053f5f828e975949687acdf0