Mirror container images for supported topics
If "--registry <...>" is passed on the command-line or in config file or
in DCI_REGISTRY environment variable, dci-downloader will mirror the
container images for the components having some.

This requires a working anonymous registry service running at the specified URL.

This implementation will only work if skopeo with the sync subcommand is
available (>= 0.1.41) and will display a warning if skopeo is not found
or an incompatible skopeo is installed.

Change-Id: I97c1e764dc46d2bd061fe490abb21729680f25cf