Add timeouts to all requests.get() calls
Adding timeouts to requests.get() calls allows the request to fail after
a set timeout instead of being stuck for ages in case of something going
wrong in the middle of the pipe (bad behaving firewalls, etc. …).

Allow 5 seconds to connect and 10 seconds of inactivity once the
connection is etablished.

Also add retries to all methods calling requests.get().

Change-Id: I26b45909b3f1bc1daac7d48296736630e2c1e2e5